Rename package to common.domain.resourcecontainer.projectimportexport
