--- common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/resourcecontainer/export/ExportResult.kt
@@ -1,6 +1,0 @@
-package org.wycliffeassociates.otter.common.domain.resourcecontainer.export
-
-enum class ExportResult {
-    SUCCESS,
-    FAILURE
-}
