+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/resourcecontainer/projectimportexport/ExportResult.kt
@@ -1,0 +1,6 @@
+package org.wycliffeassociates.otter.common.domain.resourcecontainer.projectimportexport
+
+enum class ExportResult {
+    SUCCESS,
+    FAILURE
+}
