+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/persistence/DirectoryProvider.kt
@@ -3,7 +3,7 @@ package org.wycliffeassociates.otter.jvm.workbookapp.persistence
 import org.wycliffeassociates.otter.common.data.model.Collection
 import org.wycliffeassociates.otter.common.data.model.ResourceMetadata
 import org.wycliffeassociates.otter.common.persistence.IDirectoryProvider
-import org.wycliffeassociates.otter.jvm.workbookapp.domain.resourcecontainer.export.NioZipFileWriter
+import org.wycliffeassociates.otter.jvm.workbookapp.domain.resourcecontainer.projectimportexport.NioZipFileWriter
 import org.wycliffeassociates.resourcecontainer.ResourceContainer
 import java.io.File
 import java.nio.file.FileSystems
